+++ app/components/MapMobile.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import { useState, useMemo } from "react";
+import { CORE, RUBI, PETALS, RUBI_LINKS, SUBS } from "./mapData";
+
+// Mobile-specific helpers (can reuse polar, fanPositions if needed)
+export default function MapMobile() {
+  const [openPetal, setOpenPetal] = useState<string|null>(null);
+  const [focusedSubKey, setFocusedSubKey] = useState<string|null>(null);
+
+  // ...mobile-specific layout logic, or a vertical list, or a simplified SVG...
+
+  return (
+    <div className="w-full flex flex-col items-center">
+      {/* Example: vertical list for mobile */}
+      <div className="text-center text-slate-400">[Mobile Map/List goes here]</div>
+    </div>
+  );
+}
