+++ rubi-map/app/components/index.ts
@@ -1,0 +1,4 @@
+// This file exports the components from MapDesktop and MapMobile for easy import in other parts of the application.
+
+export { default as MapDesktop } from './MapDesktop';
+export { default as MapMobile } from './MapMobile';
