+++ rubi-map/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "rubi-map",
+  "version": "1.0.0",
+  "description": "A project for visualizing the Retroactive Universal Basic Income policy map with responsive design for desktop and mobile views.",
+  "main": "app/page.tsx",
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
+  },
+  "keywords": [
+    "react",
+    "nextjs",
+    "typescript",
+    "responsive",
+    "map"
+  ],
+  "author": "Your Name",
+  "license": "MIT",
+  "dependencies": {
+    "framer-motion": "^4.1.17",
+    "lucide-react": "^0.1.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "next": "^12.0.7"
+  },
+  "devDependencies": {
+    "typescript": "^4.5.4",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.9"
+  }
+}
